requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.4.0
replicate==0.29.0
requests==2.31.0
numpy==1.24.3